savepoint

+++ Sources/Auth/Services/Common/DisconnectPairService.swift
@@ -1,11 +1,15 @@
 
 import Foundation
 import WalletConnectNetworking
+import JSONRPC
 import WalletConnectKMS
 import WalletConnectUtils
 import WalletConnectPairing
 
 class DisconnectPairService {
+    enum Errors: Error {
+        case pairingNotFound
+    }
     private let networkingInteractor: NetworkInteracting
     private let kms: KeyManagementServiceProtocol
     private let pairingStorage: WCPairingStorage
@@ -18,14 +22,46 @@ class DisconnectPairService {
         self.networkingInteractor = networkingInteractor
         self.kms = kms
         self.pairingStorage = pairingStorage
+        self.logger = logger
     }
 
-    func disconnect(topic: String) async throws {
+    func delete(topic: String) async throws {
+        guard pairingStorage.hasPairing(forTopic: topic) else { throw Errors.pairingNotFound}
         let reason = AuthError.userDisconnected
         logger.debug("Will delete pairing for reason: message: \(reason.message) code: \(reason.code)")
-        try await networkingInteractor.request(<#T##RPCRequest#>, topic: <#T##String#>, tag: <#T##Int#>, envelopeType: <#T##Envelope.EnvelopeType#>)
+        let request = RPCRequest(method: AuthProtocolMethods.pairingDelete.rawValue, params: reason)
+        try await networkingInteractor.request(request, topic: topic, tag: AuthProtocolMethods.pairingDelete.requestTag)
         pairingStorage.delete(topic: topic)
         kms.deleteSymmetricKey(for: topic)
         networkingInteractor.unsubscribe(topic: topic)
+    }
+}
+
+
+enum AuthProtocolMethods: String {
+    case authRequest = "wc_authRequest"
+    case pairingDelete = "wc_pairingDelete"
+    case pairingPing = "wc_pairingPing"
+
+    var requestTag: Int {
+        switch self {
+        case .authRequest:
+            return 3000
+        case .pairingDelete:
+            return 1000
+        case .pairingPing:
+            return 1002
+        }
+    }
+
+    var responseTag: Int {
+        switch self {
+        case .authRequest:
+            return 3001
+        case .pairingDelete:
+            return 1001
+        case .pairingPing:
+            return 1003
+        }
     }
 }
